engine: extract esm/space_time_controller.js
